+ refactor `query_results_template` module
